Tested Basic Lighting Closes #13
Tested basic lighting by adding a PointLIght2D with the texture being a light circle. I attached this to the player so that it follows the player when they move.  I also added the camera as a child node to the player so that the camera follows the player instead of being situated in one place. I added a CanvasModulate and set it to a black color to make the scene darker to display the light around the player. I also added basic collision so that you can see other objects in the scene when the player is moving. Final I organized all the files we had into separate folders to be organized more neatly.