Enhanced the Environment Closes #56
I changed the gray background tiles to basic grass tiles. Added two different types of trees and a type of bush. I made it so that each of these has collisions as well as Y sort capability (where if the player is behind the object it will actual put the sprite behind the object and vice versa with the front). Also made the character a bit smaller to fit the size of the trees and other things.