Added Thorn Traps Closes #67
Added three different thorn images to the project.  Added thorns to a new thorns tilemap. Added a new scene that is a area2d that accounts for tiles that will do damage. Turned on y-sort for the player. Moved the resetting of coins to to the on_reload_timeout function for when the player dies.  Fixed the rats movement speed and path range.