Better approach to the shop Closes #125
Previously when entering the shop it took you to a whole new scene and when you click the back button it would go to the old scene and put you in the position you were last at. The problem with this was that all the enemies and other things would reload to. So what I changed was just to have the game pause and just display the shop scene over the game that way everything stays the same whenever you return from the shop.